+++ test/test_wcstombs.c
@@ -10,36 +10,36 @@ START_TEST(test_wcstombs)
     const wchar_t  wchars[] = { L'h', L'e', L'l', L'l', L'o', L'\0' };
     const wchar_t  wchars_bad[] = { L'h', L'i', 0xffff, 0 };
     const char     cchars[] = "hello";
-    int            ret;
+    size_t         ret;
     char bytes[BUFSIZ];
 
     // Given a NULL destination, these functions count valid characters.
     errno = 0;
     ret = _wcstombs(NULL, wchars, 0);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5);
+    ck_assert_int_eq(ret, 5U);
 
     errno = 0;
     ret = _wcstombs(NULL, wchars, 4);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5);
+    ck_assert_int_eq(ret, 5U);
 
     errno = 0;
     ret = _wcstombs(NULL, wchars, 256);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5);
+    ck_assert_int_eq(ret, 5U);
 
     errno = 0;
     memset(bytes, 'x', sizeof(bytes));
     ret = _wcstombs(bytes, wchars, 0);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, -1);
+    ck_assert_int_eq(ret, (size_t)-1);
 
     errno = 0;
     memset(bytes, 'x', sizeof(bytes));
     ret = _wcstombs(bytes, wchars, 4);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 4);
+    ck_assert_int_eq(ret, 4U);
     bytes[5] = 0;
     ck_assert_str_eq("hellx", bytes);
 
@@ -47,7 +47,7 @@ START_TEST(test_wcstombs)
     memset(bytes, 'x', sizeof(bytes));
     ret = _wcstombs(bytes, wchars, 256);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5);
+    ck_assert_int_eq(ret, 5U);
     bytes[5] = 0;
     ck_assert_str_eq(cchars, bytes);
 
@@ -55,7 +55,7 @@ START_TEST(test_wcstombs)
     memset(bytes, 'x', sizeof(bytes));
     ret = _wcstombs(bytes, wchars, 6);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5);
+    ck_assert_int_eq(ret, 5U);
     bytes[5] = 0;
     ck_assert_str_eq(cchars, bytes);
 
@@ -64,18 +64,18 @@ START_TEST(test_wcstombs)
     errno = 0;
     ret = _wcstombs(NULL, wchars_bad, 0);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5); // TODO: return == -1 good ?
+    ck_assert_int_eq(ret, 5U); // TODO: return == -1 good ?
 
     errno = 0;
     ret = _wcstombs(NULL, wchars_bad, 256);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5); // TODO: return == -1 good ?
+    ck_assert_int_eq(ret, 5U); // TODO: return == -1 good ?
 
     memset(bytes, 'x', sizeof(bytes));
     errno = 0;
     ret = _wcstombs(bytes, wchars_bad, 256);
     ck_assert(errno != EILSEQ);
-    ck_assert_int_eq(ret, 5); // TODO: return == -1 good ?
+    ck_assert_int_eq(ret, 5U); // TODO: return == -1 good ?
     ck_assert(bytes[0] == L'h');
     ck_assert(bytes[1] == L'i');
     ck_assert(bytes[6] == 'x'); // TODO: bytes 2,3,4,5 is broken
